feat：项目级别的变量语法风格设置增加风险提示和风险确认 #11407

+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/user/UserPipelineResource.kt
@@ -644,7 +644,7 @@ interface UserPipelineResource {
         @Parameter(description = "流水线方言")
         @QueryParam("dialect")
         dialect: PipelineDialectType
-    ): Result<Int>
+    ): Result<Long>
 
     @Operation(summary = "获取流水线列表根据流水线方言")
     @GET
