+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/UserPipelineResourceImpl.kt
@@ -780,7 +780,7 @@ class UserPipelineResourceImpl @Autowired constructor(
         userId: String,
         projectId: String,
         dialect: PipelineDialectType
-    ): Result<Int> {
+    ): Result<Long> {
         return Result(
             pipelineListFacadeService.countPipelineByDialect(
                 projectId = projectId, dialect = dialect
